Re-organize tests
