+++ automation-test/deleteUser.test.ts
@@ -1,5 +1,5 @@
 const expect = require("chai").expect;
-import { setUrl } from "../test/testData/set-url";
+import { setUrl } from "./testData/setUrl";
 import Common from "./commonMethods";
 import UserData from "./testData/userData";
 const url = setUrl();
